refactor(quote request form): refactors component so that multiple layouts can be used with same form logic
