+++ app/components/Home.tsx
@@ -41,7 +41,7 @@ export default function Home(props: {
             return <CompanyValues key={i} {...block} />;
           }
           case "PageHomeBlocksQuoteSection": {
-            return <RequestQuote key={i} {...block} />;
+            return <RequestQuote key={i} {...block} layoutType="homePage" />;
           }
           case "PageHomeBlocksLocation": {
             return <Location key={i} {...block} />;
